Rename canopen_mock_slave to canopen_fake_slaves
