+++ canopen_fake_slaves/config/cia402_slave.eds
