+++ canopen_fake_slaves/config/simple_slave.eds
