+++ canopen_fake_slaves/package.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0"?>
 <?xml-model href="http://download.ros.org/schema/package_format3.xsd" schematypens="http://www.w3.org/2001/XMLSchema"?>
 <package format="3">
-  <name>canopen_mock_slave</name>
+  <name>canopen_fake_slaves</name>
   <version>0.0.0</version>
   <description>Package with mock canopen slave</description>
   <maintainer email="cmh@ipa.fraunhofer.de">Christoph Hellmann Santos</maintainer>
