+++ canopen_ros2_control/launch/canopen_system.launch.py
@@ -139,7 +139,7 @@ def launch_setup(context, *args, **kwargs):
     )
 
     slave_launch = PathJoinSubstitution(
-        [FindPackageShare("canopen_mock_slave"), "launch", "basic_slave.launch.py"]
+        [FindPackageShare("canopen_fake_slaves"), "launch", "basic_slave.launch.py"]
     )
     slave_node_1 = IncludeLaunchDescription(
         PythonLaunchDescriptionSource(slave_launch),
