+++ canopen_tests/CMakeLists.txt
@@ -31,15 +31,9 @@ install(DIRECTORY
 
 
 if(BUILD_TESTING)
-  find_package(ament_lint_auto REQUIRED)
-  # the following line skips the linter which checks for copyrights
-  # comment the line when a copyright and license is added to all source files
-  set(ament_cmake_copyright_FOUND TRUE)
-  # the following line skips cpplint (only works in a git repo)
-  # comment the line when this package is in a git repo and when
-  # a copyright and license is added to all source files
-  set(ament_cmake_cpplint_FOUND TRUE)
-  ament_lint_auto_find_test_dependencies()
+  find_package(launch_testing_ament_cmake REQUIRED)
+  add_launch_test(launch_tests/test_proxy_driver.py)
+  add_launch_test(launch_tests/test_proxy_lifecycle_driver.py)
 endif()
 
 ament_package()
