+++ canopen_tests/launch/proxy_setup.launch.py
@@ -14,7 +14,7 @@ def generate_launch_description():
         PythonLaunchDescriptionSource(
             [
                 os.path.join(
-                    get_package_share_directory("canopen_mock_slave"), "launch"
+                    get_package_share_directory("canopen_fake_slaves"), "launch"
                 ),
                 "/basic_slave.launch.py",
             ]
@@ -30,7 +30,7 @@ def generate_launch_description():
         PythonLaunchDescriptionSource(
             [
                 os.path.join(
-                    get_package_share_directory("canopen_mock_slave"), "launch"
+                    get_package_share_directory("canopen_fake_slaves"), "launch"
                 ),
                 "/basic_slave.launch.py",
             ]
